+++ src/calendar/dot/index.js
@@ -11,6 +11,10 @@ const Dot = ({theme, isMarked, isDisabled, dotColor, isToday, isSelected}) => {
   if (isMarked) {
     dotStyle.push(style.visibleDot);
 
+    if (isToday) {
+      dotStyle.push(style.todayDot);
+    }
+
     if (isDisabled) {
       dotStyle.push(style.disabledDot);
     }
@@ -21,10 +25,6 @@ const Dot = ({theme, isMarked, isDisabled, dotColor, isToday, isSelected}) => {
 
     if (dotColor) {
       dotStyle.push({backgroundColor: dotColor});
-    }
-
-    if (isToday) {
-      dotStyle.push(style.todayDot);
     }
   }