+++ src/expandableCalendar/style.js
@@ -1,4 +1,4 @@
-import {StyleSheet, Platform} from 'react-native';
+import {StyleSheet, Platform, I18nManager} from 'react-native';
 import * as defaultStyle from '../style';
 
 
@@ -51,6 +51,7 @@ export default function styleConstructor(theme = {}) {
       paddingTop: 24, // 8
       paddingBottom: 8,
       paddingLeft: 20,
+      paddingRight: 20,
       backgroundColor: appStyle.calendarBackground,
       textAlign: 'left',
       textTransform: 'uppercase'
@@ -102,7 +103,8 @@ export default function styleConstructor(theme = {}) {
       color: appStyle.textSectionTitleColor
     },
     arrowImage: {
-      tintColor: appStyle.arrowColor
+      tintColor: appStyle.arrowColor,
+      transform: I18nManager.isRTL ? [{scaleX: -1}] : undefined
     },
     todayButtonContainer: {
       alignItems: appStyle.todayButtonPosition === 'right' ? 'flex-end' : 'flex-start',